requirements.txt
Django>=4.2.0,<5.0.0
django-storages>=1.13.0
boto3>=1.26.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
dj-database-url>=1.0.0
gunicorn>=20.1.0
celery>=5.3.0
django-celery-beat>=2.5.0
redis>=4.6.0